Update soundscrape.py
I added a upper length limit argument to filter out mixes when downloading from soundcloud as per issue #175. This is my first attempt to contribute to something on github so im not exactly sure if theres anything else i need to do to test it but it works great for me. as of now it lets you enter "-limit 5" for example, however entering 5:00 wouldnt work, and 5.5 would be 5:30, although for the purpose of filtering mixes, it gets the job done